+++ actiontext/CHANGELOG.md
@@ -1,3 +1,14 @@
+*   Treat `Content#to_html` and `Content#to_trix_html` as safe
+
+    ```ruby
+    content = ActionText::Content.new("<div>hello world</div>")
+
+    content.to_html.html_safe? # => true
+    content.to_trix_html.html_safe? # => true
+    ```
+
+    *Sean Doyle*
+
 *   Compile ESM package that can be used directly in the browser as actiontext.esm.js
 
     *Matias Grunberg*